implementation of file appending logic

+++ roboclimate/weather_spider.py
@@ -5,6 +5,7 @@ import requests
 from requests.exceptions import ConnectionError
 from tenacity import retry, retry_if_exception_type, wait_fixed, stop_after_attempt
 import boto3
+import botocore
 
 # type alias
 csv_row = "list[str]"
@@ -25,7 +26,7 @@ CITIES = {"london": 2643743,
 WEATHER_RESOURCE = "weather"
 S3_BUCKET_NAME = "roboclimate"
 S3_OBJECT_PREFIX = "weather"
-CSV_HEADER = ['temp', 'pressure', 'humidity', 'wind_speed', 'wind_deg', 'dt', 'today']
+CSV_HEADER = "temp,pressure,humidity,wind_speed,wind_deg,dt,today"
 TOLERANCE = {'positive_tolerance': 1200, 'negative_tolerance': 60}  # tolerance in seconds
 
 # global variables
@@ -160,22 +161,34 @@ def transform_data(weather_data: requests.Response, utcnow_date_f, tolerance: "d
         raise ex
 
 
-def write_to_filesystem(file_name, data):
+def write_to_filesystem(file_name: str, data: str):
     logger.info(f'writing file {file_name}')
-    with open(f"{file_name}", 'w', encoding='UTF-8') as f:
-        f.write(data)
+    if os.path.exists(file_name):
+        # append data
+        with open(f"{file_name}", 'a', encoding='UTF-8') as f:
+            f.write(data)
+    else:
+        # add header when file is first created
+        with open(f"{file_name}", 'w', encoding='UTF-8') as f:
+            f.write(f"{CSV_HEADER}\n{data}")
 
 
-def write_to_s3(file_name, data):
-    logger.info(f'writing object {file_name} to bucket {S3_BUCKET_NAME}')
+def write_to_s3(file_name: str, data: str):
+    try:
+        existing_data = s3.get_object(Bucket=S3_BUCKET_NAME, Key=file_name)['Body'].read().decode('UTF-8')
+    except botocore.exceptions.ClientError:
+        # add header when file is first created
+        data = f"{CSV_HEADER}\n{data}"
+    else:
+        # append data
+        data = f"{existing_data}{data}"
+    logger.info(f'writing object {file_name} in bucket {S3_BUCKET_NAME}')
     s3.put_object(Body=data, Bucket=S3_BUCKET_NAME, Key=file_name)
 
 
 def write_data(city_name: str, weather_data_csv: csv_rows, write_f, s3_object_prefix):
     csv_file_name = f"{s3_object_prefix}/{WEATHER_RESOURCE}_{city_name}.csv"
-    weather_data_csv_with_header = [CSV_HEADER]
-    weather_data_csv_with_header.extend(weather_data_csv)
-    csv_data_serialized = '\n'.join([','.join(map(str, row)) for row in weather_data_csv_with_header])
+    csv_data_serialized = ('\n'.join([','.join(map(str, row)) for row in weather_data_csv]))+'\n'
     write_f(csv_file_name, csv_data_serialized)
